Moves tangential tests into another file.
