+++ pkg/printers/internalversion/BUILD
@@ -9,6 +9,7 @@ load(
 go_test(
     name = "go_default_test",
     srcs = [
+        "additional_printers_test.go",
         "printers_test.go",
         "sorted_resource_name_list_test.go",
     ],
@@ -41,8 +42,8 @@ go_test(
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/cli-runtime/pkg/genericclioptions:go_default_library",
         "//staging/src/k8s.io/cli-runtime/pkg/printers:go_default_library",
+        "//vendor/gopkg.in/yaml.v2:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
-        "//vendor/sigs.k8s.io/yaml:go_default_library",
     ],
 )
